+++ modules/sdk-core/src/bitgo/index.ts
@@ -15,7 +15,7 @@ export * from './inscriptionBuilder';
 export * from './internal';
 export * from './keychain';
 export * as bitcoin from './legacyBitcoin';
-export * from './lightning';
+export * from './lightning/custodial';
 export * from './market';
 export * from './pendingApproval';
 export * from './recovery';
