+++ modules/sdk-core/src/bitgo/lightning/index.ts
@@ -1,6 +1,2 @@
-export * from './iLightning';
-export * from './lightning';
-export * from './lnurlCodec';
-export * from './lightningUtils';
 export * from './signableJson';
 export * from './signature';
